# src/main/java/dk/viplev/api/adapter/inbound/rest/EnvironmentApiDelegateImpl.java
package dk.viplev.api.adapter.inbound.rest;

import dk.viplev.api.adapter.inbound.rest.dto.EnvironmentDTO;
import dk.viplev.api.adapter.inbound.rest.dto.ServiceDTO;
import dk.viplev.api.port.inbound.EnvironmentService;
import dk.viplev.api.port.inbound.ServiceService;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Component;

import java.util.List;
import java.util.UUID;

@Component
@RequiredArgsConstructor
@Slf4j
public class EnvironmentApiDelegateImpl implements EnvironmentApiDelegate {

    private final EnvironmentService environmentService;
    private final ServiceService serviceService;

    @Override
    public ResponseEntity<List<EnvironmentDTO>> listEnvironments() {
        return ResponseEntity.ok(environmentService.listEnvironments());
    }

    @Override
    public ResponseEntity<EnvironmentDTO> createEnvironment(EnvironmentDTO environmentDTO) {
        EnvironmentDTO created = environmentService.createEnvironment(environmentDTO);
        return ResponseEntity.status(org.springframework.http.HttpStatus.CREATED).body(created);
    }

    @Override
    public ResponseEntity<EnvironmentDTO> getEnvironment(UUID environmentId) {
        return ResponseEntity.ok(environmentService.getEnvironment(environmentId));
    }

    @Override
    public ResponseEntity<EnvironmentDTO> updateEnvironment(UUID environmentId, EnvironmentDTO environmentDTO) {
        return ResponseEntity.ok(environmentService.updateEnvironment(environmentId, environmentDTO));
    }

    @Override
    public ResponseEntity<Void> deleteEnvironment(UUID environmentId) {
        environmentService.deleteEnvironment(environmentId);
        return ResponseEntity.noContent().build();
    }

    @Override
    public ResponseEntity<List<ServiceDTO>> listServices(UUID environmentId) {
        return ResponseEntity.ok(serviceService.listServices(environmentId));
    }

    @Override
    public ResponseEntity<ServiceDTO> getService(UUID environmentId, UUID serviceId) {
        return ResponseEntity.ok(serviceService.getService(environmentId, serviceId));
    }
}