# src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VIPLEV
  description: This is an API specification
  version: 1.0.0
security:
  - bearerAuth: []
...
paths:
...
  /v1/environments:
    get:
      tags:
        - Environment
      summary: Get all environments
      operationId: listEnvironments
      responses:
        '200':
          description: List of all environments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnvironmentDTO'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '500':
          $ref: '#/components/responses/500InternalServerError'
    post:
      tags:
        - Environment
      summary: Create environment
      operationId: createEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentDTO'
      responses:
        '201':
          description: Environment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentDTO'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  /v1/environments/{environmentId}:
    get:
      tags:
        - Environment
      summary: Get an environment
      operationId: getEnvironment
      parameters:
        - name: environmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Environment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentDTO'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
    put:
      tags:
        - Environment
      summary: Update an environment
      operationId: updateEnvironment
      parameters:
        - name: environmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentDTO'
      responses:
        '200':
          description: Environment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentDTO'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
    delete:
      tags:
        - Environment
      summary: Delete an environment
      operationId: deleteEnvironment
      parameters:
        - name: environmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Environment deleted
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  /v1/environments/{environmentId}/services:
    parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Environment
      summary: Get all services on an Environment
      operationId: listServices
      responses:
        '200':
          description: List of all services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceDTO'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
    post:
      tags:
        - Agent
      summary: Agent can post a list of services
      operationId: registerServices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRegistrationDTO'
      responses:
        '201':
          description: Services created
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  /v1/environments/{environmentId}/services/{serviceId}:
    parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: serviceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Environment
      summary: Get a single service
      operationId: getService
      responses:
        '200':
          description: Service found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDTO'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    400BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
          examples:
            invalidField:
              summary: Invalid field
              value:
                type: /errors/bad-request
                title: Bad Request
                status: 400
                detail: The field 'name' is required
                instance: /v1/environments
            invalidFormat:
              summary: Invalid format
              value:
                type: /errors/bad-request
                title: Bad Request
                status: 400
                detail: The field 'port' must be a number
                instance: /v1/environments
    401Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
          examples:
            unauthorized:
              summary: Missing or invalid token
              value:
                type: /errors/unauthorized
                title: Unauthorized
                status: 401
                detail: Authentication token is missing or invalid
                instance: /v1/environments
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
          examples:
            notFound:
              summary: Resource not found
              value:
                type: /errors/not-found
                title: Not Found
                status: 404
                detail: Resource with id '123' not found
                instance: /v1/environments/123
    500InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
          examples:
            serverError:
              summary: Server error
              value:
                type: /errors/internal-server-error
                title: Internal Server Error
                status: 500
                detail: An unexpected error occurred while processing the request
                instance: /v1/environments
  schemas:
    EnvironmentDTO:
      type: object
      required:
        - name
        - type
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: Unique ID for environment
          examples:
            - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        name:
          type: string
          description: Name of environment
        description:
          type: string
          description: Description of environment
        type:
          type: string
          enum: [docker, kubernetes]
          description: Container platform type
        token:
          type: string
          readOnly: true
          description: Authentication token for the agent, generated by the system
        agentCommand:
          type: string
          readOnly: true
          description: Generated command to spin up the agent (e.g. docker run or kubectl apply)
        agentLastSeenAt:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the agent last polled the server
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when environment was created
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when environment was last updated
...
    HostDTO:
      type: object
      required:
        - name
        - machineId
        - os
        - ipAddress
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: Unique ID for host
          examples:
            - "c3d4e5f6-a7b8-9012-cdef-123456789012"
        name:
          type: string
          description: Hostname
        machineId:
          type: string
          description: Unique machine identifier (e.g. /etc/machine-id on Linux)
        ipAddress:
          type: string
          description: IP address of the host
        os:
          type: string
          description: Operating system name
        osVersion:
          type: string
          description: Operating system version
        cpuModel:
          type: string
          description: CPU model name
        cpuCores:
          type: integer
          description: Number of physical CPU cores
        cpuThreads:
          type: integer
          description: Number of CPU threads
        ramTotalBytes:
          type: integer
          format: int64
          description: Total RAM in bytes
        ramSpeedMhz:
          type: integer
          description: RAM speed in MHz
        ramType:
          type: string
          description: RAM type (e.g. DDR4, DDR5)
    ServiceRegistrationDTO:
      type: object
      required:
        - host
        - services
      properties:
        host:
          $ref: '#/components/schemas/HostDTO'
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceDTO'
    ServiceDTO:
      type: object
      required:
        - serviceName
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: Unique ID for service
          examples:
            - "d4e5f6a7-b8c9-0123-def0-1234567890ab"
        serviceName:
          type: string
          description: Stable service name (e.g. mystack_nginx.1)
        imageSha:
          type: string
          description: SHA hash of the container image
        imageName:
          type: string
          description: Name of the container image
        cpuLimit:
          type: number
          format: double
          description: CPU limit allocated to the service
        cpuReservation:
          type: number
          format: double
          description: CPU reservation for the service
        memoryLimitBytes:
          type: integer
          format: int64
          description: Memory limit in bytes
        memoryReservationBytes:
          type: integer
          format: int64
          description: Memory reservation in bytes
...
    ErrorDTO:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          examples:
            - /errors/bad-request
        title:
          type: string
          description: A short, human-readable description of the problem type
          examples:
            - Bad Request
        status:
          type: integer
          description: HTTP status code
          examples:
            - 400
        detail:
          type: string
          description: A detailed explanation of the problem
          examples:
            - The field 'name' is required
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence of the problem
          examples:
            - /v1/environments/123
...